Clean up items file
Move help types and functions to make it easier to see which items
exists.
